[kda-mtp] simplify fused conv KVBuffer path
